refactor(connector): unify extracing table schema from avro schema
